Fix for postgresql-libpg and musl (fixes #948)
This combined with setting the `use-pkg-config` flag
on `postgresql-libpq` should get it building.

Add the following to your `cabal.project` file or
to `cabalProjectLocal`.

```
package postgresql-libpq
  flags: +use-pkg-config
```

This PR also adds a "hackage quirk" for postgrest
to add a `cabalProject` to set the flag for that
package (when built as a `tool` or with `hackage-package`).

Unfortunately we do not have a good way to default
flags in a way that is picked up before `cabal configure`
runs.