Add hackage and stackage as flake inputs
This will mean that the pinned commits in `hackage-src.json` and `stackage-src.json` will not be used for flakes.  Instead the nix flake system will be free to choose the most recent commit.

One possible drawback of this system is that if we wanted to roll back to an old commit of `haskell.nix` and also pin `hackage` and `stackage` to whatever was current at the time we would have to look up the commit manually (it will not be possible to use `follows` to get the appropriate commits).

If we create a new `flake.lock` for a project using the latest haskell.nix though it get pins for `hackage` and `stackage` at the time it is created.  So if we keep the `flake.lock` in the projects repo we can easily roll back to it.

Fixes #1132