Simplify the way evalPackages works (#1162)
This change uses the same overlay for both flakes and non flake projects.  It chooses where evalPackages should run based on `builtins ? currentSystem` (`currentSystem` will not exist when building flake).

This will be needed to make sure the `evalPackages` still works correctly in #1151 (otherwise the use of `flake-compat` will cause eval-on-build to be used).