Changes for source-repository-package handling (#107)
Changes needed to support improved handling of `source-repository-package` in `cabal.project`.  Currently the `haskell.nix` 
`cabalProject` functions replace these with regular `packages:` pointing at the downloaded source and these are not treated the same by `cabal`.  We are planning to change it so:

* `haskell.nix` will use `git` to create local repos and a replacement `source-repoository-package`.
* Once the `plan.json` is generated and `plan-to-nix ` has replace these the `/nix/store` paths (so that materialization can work).
*  Have the src mapped to `fetchgit` in haskell.nix.

For this to work we need to be able to identify the `source-repository-packages`.  This change added `url`, `rev` and `sha256` attributes to the `src` for `source-repository-packages`.

It also fixes an issue where the `source-repository-package` is expected to be in `hackage` and allows fixes a problem with the cache if a `source-repository-package` did not have a sha256 provided (not needed if it is coming from `/nix/store`).

Used by https://github.com/input-output-hk/haskell.nix/pull/1166