SPDX: Provide default for isFsfLibre
The `isFsfLibre` property isn't always included, as opposed to the
`isOsiApproved` property. To avoid errors we default it to `false` if
it's missing.

I opted for the simpler, more verbose way to write this.
`lic.isFsfLibre or isOsiApproved` would work as well and reads well but
it's not immediately clear that we're testing whether the field exists
and defaulting rather than doing a boolean OR.