Better error when GHC is too old to build
Currently we often get confusing errors (infinite loops looking for impossible materialization or GHC build errors) when trying to build a GHC version that is never going to work on the platform we are building for.  In particular it is common to on M1 Macs to forget to include `--system x86_64-darwin` and wind up with a confusing error.

With this change the error is more clear:

```
hamish@Hamishs-MBP haskell.nix % nix-build -A pkgs-unstable.haskell-nix.compiler.ghc8105
error: Desired GHC (8.10.5) is older than the bootstrap GHC (8.10.7) for this platform (aarch64-apple-darwin).
```