Look for DLLs in /bin directory of pkgconfig deps (#1422)
We include the /bin directory of `component.libs` in `extra-lib-dirs` on windows, but not `components.pkgconfig`.  This causes confusing differences between the behavior depending on how a dependency was specified in the `.cabal` file or added in a module.  If it wounds up in `components.libs` the need for DLLs would propagate correctly, but if it was only in `componets.pkgconfig` DLLs would not be included when building TH code and did not get symlinked to `exe` component output /bin dirs.