Fix cleanGit when src is a git submodule.
Suppose you have a repository with a submodule checked out in the
subdirectory `smod`.

`cleanGit { src = ./smod; }` would return the following error:

    error: getting status of '/.git/modules/smod': No such file or directory

`smod/.git` is a file that contains a reference to the git module repo.

    gitdir: ../.git/modules/smod

This is the same as a worktree (in fact, isWorktree = true here),
but it's a relative path instead of an absolute path.