Workaround possible TH environment overflow running `wine`
We still do not understand the cause, but wine seems to fail to start (we get a segfault) when building `cardano-wallet-cli` test `unit`.  It seems to be an overflow somewhere in bash or very early in the `wine` startup code (since nothing is logged even with `WINEDEBUG=+all`).

This changes `unsets` the large `pkgsHostTargetAsString` environment variable before wine is started (after that variable has been used).  This seems to be enough to avoid the issue though it is worrying that it is not clear why.