Let cabal configure errors break IFD (#1744)
In the past hydra didn't give good feedback on eval time errors.  Deferring the errors was a way to get them into a build log (instead of the eval log that would get overwritten with the next eval).

I'm not sure how well this actually worked since it is super easy to introduce an eval time dependency on a broken project.  For instance if you try to enumerate the components in a project haskell that failed to configure the error would be forced (in some cases a more cryptic one).

Now that hydra is better at reporting eval errors I think we should try to fail fast again.