Fix aarch64-linker, take 2 (#1791)
This one moves most of the allocation logic into the m32 allocator, while also completely ignoring the lower requirement. This of course is not ideal, however we've been mapping pages at the oder of 100x (4096 vs 40) the required space, and thus often ran out of memory due to sections per function.

This is now much improved. It also revealed at least one out-of-bounds memory write issue.