Discard string context when using antiquoted string as attrset key (#1910)
This manifested in practice because I was trying to do something like
this:
```
repository cardano-haskell-packages:
  url: file:${CHaP}
  secure: True
```
which gave me errors about not being allowed to have string context. We
do actually need the string context (otherwise we end up not depending
on the repository, and the sandbox prevents us from looking at it!), so
we have to be careful about where to discard. The simplest thing is to
discard where it's needed: as we create the attrset key.