Adding some note about local vs. global nix.conf
I faced some puzzling behaviour where artifacts expected to be cached were not found and nix built them. It turned out the local configuration was wrong. It's not clear how this configuration was written in the first place as the most likely candidate I checked should have had different caches.