Add exceptions and stm to nonReinstallablePkgs for ghc901
We did not merge #1183 because we were concerned that it would change the  exceptions and stm used for users that did not want to build GHC. However `reinstallableLibGhc = true` is now the default and most projects should not be setting it to `false`.  That means this change will now only affect projects that likely want to use the built in `ghc` package.