Remove `configFiles` and `ghc` component outputs (#1940)
* Remove `configFiles` and `ghc` component outputs

These outputs were added when the derivations that used to provide them were removed (they instead the code that creates them is run in the `configurePhase`).

These outputs were added when the derivations that used to provide them were removed (they instead the code that creates them is run in the `configurePhase`).

We think we might be better off without theses. They have two drawbacks:

* When running `genericBuild` inside a shell they must be created manually (with `export configFile=$(mktemp -d)` and `export ghc=$(mktemp -d)`).

* The stop `disallowedReferences` from working (see #1934).

* It is probably a non zero overhead to including them in the nix cache.

This change also adds `keepConfigFiles` and `keepGhc` that can be used like `keepSource` to allow references to `configFiles` and `ghc` to persist.  This is needed by `ghc-paths` and `cabal-doctests` packages.