Update static overrides, to always use final
We want to make sure that the static overrides happen last. That is if someone decided to swap out a library, we still want to apply the static logic at the final step.

This came up in https://github.com/input-output-hk/cardano-db-sync/pull/1403/commits/a7cd5db34782e70b0044ecb97c70654506690dc5, where we imported an overlay that modified the crypto libraries _after_ the haskell.nix overlays, and thus all of static cross broke, as the static changes got overridden by the later crypto-library changes. To fix this we had to introduce order in the sequence we apply overlays in. That however is highly error prone.