Add std++ to postgres as well as setting LC_CTYPE
The reason that necessitates setting LC_CTYPE is not fully clear to me,
I've left some comments on the nixos/nixpkgs issue tracker.
It appears as if we are pulling LC_CTYPE to be some UTC8 value if it's
not set, instead of inheriting it from LC_ALL as it supposed to.

https://github.com/NixOS/nixpkgs/issues/150930#issuecomment-1553979592