Revert default value for sha256map
This breaks the behaviour when a `sha256map` is not provided, since the
logic that uses it in `cabal-project-parser` treats a `null` `sha256map`
differently from one that is present but lacks entires for a key. The
former will just try and proceed without a hash (which can succeed),
the latter always fails with a missing key error. Adding a default value
pushes us into the latter case unless we manually set the `sha256map` to
null.

I think we should try to find another way to make HLS work that doesn't
break other uses. Or we could make the `sha256map` usage logic behave
the same for a missing map and a map with a missing key.