Properly transcode absolute LoadDLL paths for WINE (#1974)
* Properly transcode absolute LoadDLL paths for WINE

We have this sub-protocol in the iserv-proxy (which follows the sub-TH
protocol). The design is horrendous, and needs to be rethought.  The
issue we were seeing is that we started the sub-protocol, because a path
looked like it was absolute, however on the receiving end (windows), the
path /foo/bar/baz, is _not_ an absolute windows path. Thus the
corresponding branch dealing with the sub-protocol was not taken and
we were stuck with messages sent for the primary protocol, and not
understood by the sub-protocol.

This change transcodes unix paths to windows paths, _IF_ the
ISERV_TARGET environment var is set to WINE. In this case we know
we can use Z:\\ as / and share the same file system. This is good enough
for haskell.nix, but a massive hack generally.

---------

Co-authored-by: Hamish Mackenzie <Hamish.K.Mackenzie@gmail.com>