Fix using /nix/store paths in cabal.project
This change makes sure paths have a context so they will be
included in the derivation inputs for the component derivations.
Without this sandbox builds cannot see the input and
fail with the error:

  do not know how to unpack source archive /nix/store/...

This PR also fixes an issue if project src has no cabal files and is
not the root directory of the repo.