Add .doc derivation to test components (#1982)
* Add a .haddock/.doc derivation to test components

Haddocks are built with --tests if we are in a test component.

* Write haddocks of tests into package prefix dir

Haddock will generate a <componentName>/ directory within this path
automatically. So a package with name "foo" and test suite "bar" will
have docdir paths:

$doc/share/doc/foo/html/index.html # library haddocks
$doc/share/doc/foo/html/bar/index.html # bar test suite haddocks

* Remove references to component derivation

Makes the cleanup step search more liberally for .html files to remove
references from. This is needed because test suite haddocks are placed
in a <package-name>/html/<test suite name>/ subdirectory.

* Fix test

---------

Co-authored-by: Hamish Mackenzie <Hamish.K.Mackenzie@gmail.com>