Allow passing a tarball to hackage2nix without a url (#2014)
cabal has started using XDG paths for the index and a quick fix is
to just pass hackage2nix the path to the tarball. But so far we are
forced to specify an url as well, which is not needed when you use
hackageTarball. This difference in behaviour causes all nix files to
change.

This change allows passing a tarball and no url, just like we do with
hackage.