Force msvcrt instead of ucrt for GHC
Nixpkgs `mingw64` is a msvcrt based toolchain. GHC 9.4+ moved on
to a ucrt based toolchain. We can hover reverse this change in GHC
and keep using msvcrt for now.

We also need to prevent the linker for linking ucrt, as that is
is incompatible with ucrt. By doing so we can also upgrade to a
newer wine, which contains ucrt libs. This would previously have
thrown our TH logic off.