Use haskell.nix in for hadrian
Using musl64 version of nix-tools goes some way to avoid recompiles of GHC when updating haskell.nix.  However any changes that impact the way `hadrian` will be built (for instance changes to `builder/comp-builder.nix`) will cause `hadrian` to need to be rebuilt.  This intern causes GHC >=9.4 to need rebuilding.

To make this less likely this change adds a `pinned-haskell-nix` using the `nix-tools` subflake.  Then hadrian is built using this.

Also when the static version of `nix-tools` cannot be used (macOS), a pinned version is used instead.

The `nixpkgs` used is also pinned, making it easier for people to use other versions of nixpkgs.

The `hackage.nix` is passed in to the `nix-tools` subflake to avoid downloading two copies when setting up a new machine (stackage.nix is not used).