Allow order independent overlays (#2043)
Currently the haskell.nix overlays ignore any `defaultModules` and `extraPkgconfigMappings` set in `prev`.  This means it is not possible to have an overly include these in an way that will work both before and after the haskell.nix overlays.

With this change an overlay can be written like this:

```
final: prev: {
  haskell-nix = prev.haskell-nix or {} // {
    extraPkgconfigMappings = prev.haskell-nix.extraPkgconfigMappings or {} // {
    ...
```

It can then be included before or after the haskell.nix overlays.