Add the attribute `options` to projects (#2052)
Previously we only exposed the config of the evaluated project
configuration modules under the name args (because effectively it is
what you could have passed to the project function).

This patch exposes also the options attribute from the configuration
evaluation. This is very helpful while troubleshooting how the final
configuration comes to be.