Don't require a `bin` directory for libs
When cross compiling for windows we include the `bin` directory of libraries as `extra-library-dirs` in the configure flags (so that any `.dll` files in the directory can be found).  Unfortunately this can cause TH code to fail if the directory does not exist.

This change will only include the `bin` directory of each library if it exists already.