Fix string contexts in cabal project parser
When both `repository` and `source-repository-packages` are included in a `cabal.project` (including `cabalProjectLocal`), the context needed for the `repository` is lost in the cabal project parser.

This means nix store path is not included in the dependencies of the plan-nix derivation and it fails with an error like this:

```
/nix/store/x-source/root.json:openBinaryFile: does not exist (No such file or directory)
```

This fix makes sure the context is correct by looking for /nix/store strings in the repository url and appending them to the context.