Allow reinstall of filepath in GHC 9.4.8
A cabal flag (`vendor-filepath`) was used to avoid the `template-haskell` dependency on `filepath` in the version of `template-haskell` (2.19.0.0) used by `ghc` 9.4.8.

```
    if flag(vendor-filepath)
      other-modules:
        System.FilePath
        System.FilePath.Posix
        System.FilePath.Windows
      hs-source-dirs: ./vendored-filepath .
      default-extensions:
        ImplicitPrelude
    else
      build-depends: filepath
      hs-source-dirs: .
```

The `dummy-ghc` code in haskell.nix does not evaluate conditionals (instead it just includes all of the dependencies.  This leads to issues where the planner is unable to come up with a plan because it thinks `filepath` cannot be reinstalled.

This fix works around the by hard coding a rule for `vendor-filepath`.