Respect defaults in `shellFor` (fixes #2306)
`shellFor` has an argument `allToolDeps` that when set to `true` tells haskell.nix to include all the tool dependencies of all the packages in the shell.  This does not work well for `stackProject` projects (since stackage includes a lot of packages).  So it is disabled by default in `modules/stack-project.nix`.

This default is currently ignored if `shellFor` is called directly.

This change ensures that `shellFor` respects the defaults and any shell arguments passed to the project. As a result, `project.shell` and `project.shellFor {}` will now behave consistently.