Improve remote store eval time (#2334)
This code used to put the .drv path into the string context for each
package. That meant we had to query the store each time one of these was
evaluated for the out path. This is pretty quick when using a non-remote
store. But the latency with the remote store really slows us down.

We should just calculate the outPath once instead, and use that
throughout.