Remove overlay/ghc.nix
It breaks the `override` function on the ghc compilers and it only serves to add "stackprotector" to `hardeningDisable`.  We can just add that to `compiler/ghc/default.nix`.

It looks like it also added "format", but that is already present.  For `hostPlatform.isAarch32` it adds "pie" (not already present), but we don't build GHC on 32bit arm.