Fix selection of dependencies for dev shells (#2367)
This patch fixes the way "non-haskell" dependencies are selected for dev shells (this includes `pre-existing` haskell packages).

Currently we only include "non-haskell" dependencies of the haskell packages that are needed by the shell (and not the the ones for the haskell packages that we plan to build in the shell).

This means if none of the haskell packages needed depend a particular "non-haskell" dependency it can be left out even if it will be needed to build the haskell packages we plan to build in the shell.

Fixes #2359 and probably #2256