wasm: disable tail-call flag for now (#2440)
As reported and confirmed from a ghc wasm backend user, tail-call
opcodes may cause ios webview crashes for some large haskell wasm
apps, and removing tail-call opcodes fixes the crash. Given webkit
only supported tail-call since 18.2, it's better to disable tail-call
flag for the time being and revisit this in the future.

This patch also removes --no-turbo-fast-api-calls from the dyld script
flags, since it was used to workaround
https://github.com/nodejs/node/issues/46777 when wasm module uses tail
calls. It's not needed when tail calls are not used and it has a
performance penalty.